Refactored relevant header.js methods into NoteUtils component
